requirements.txt
asgiref==3.7.2
blinker==1.4
certifi==2023.7.22
charset-normalizer==3.2.0
colorzero==2.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
dnspython==2.1.0
gpiozero==1.6.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
iotop==0.6
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.5.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pcp==5.0
pycairo==1.20.1
PyGObject==3.42.1
PyJWT==2.3.0
PyMySQL==1.1.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
pytz==2024.1
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
speedtest-cli==2.1.3
sqlparse==0.4.4
typing_extensions==4.10.0
unattended-upgrades==0.1
urllib3==2.0.4
wadllib==1.3.6
zipp==1.0.0